backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
django-cors-headers==4.4.0
PyJWT==2.9.0
pytz==2024.2
sqlparse==0.5.1
asgiref==3.8.1
django-filter>=24.3